// src/auth/dto/signup.dto.ts
export class SignupDto {
  fullname: string;
  preferredname: string;
  age: number;
  email: string;
  phonenumber: string;
  stateOfOrigin: string;
  username: string;
  educationLevel: string;
  trade: string;
  address: string;
  toolkit: string;
  profilePicture: string;
  amount: number;
  password: string; // <- don't forget password
}